Fix: deleting a report

+++ src/app/debug/table/table.component.ts
@@ -359,10 +359,12 @@ export class TableComponent implements OnInit, OnDestroy {
 
   deleteSelected(): void {
     const reportIds = this.helperService.getSelectedIds(this.tableSettings.reportMetadata);
-    this.httpService.deleteReport(reportIds, this.currentView.storageName).subscribe({
-      next: () => this.retrieveRecords(),
-      error: () => catchError(this.errorHandler.handleError()),
-    });
+    if (reportIds.length > 0) {
+      this.httpService.deleteReport(reportIds, this.currentView.storageName).subscribe({
+        next: () => this.retrieveRecords(),
+        error: () => catchError(this.errorHandler.handleError()),
+      });
+    }
   }
 
   selectAll(): void {
@@ -456,9 +458,9 @@ export class TableComponent implements OnInit, OnDestroy {
   openLatestReports(amount: number): void {
     this.httpService.getLatestReports(amount, this.currentView.storageName).subscribe({
       next: (data: Report[]) => {
-        data.forEach((report: any) => {
+        for (let report of data) {
           this.openReportEvent.next(report);
-        });
+        }
       },
       error: () => catchError(this.errorHandler.handleError()),
     });
@@ -496,8 +498,9 @@ export class TableComponent implements OnInit, OnDestroy {
     }
   }
 
-  uploadReports(event: any): void {
-    const file: File = event.target.files[0];
+  uploadReports(event: Event): void {
+    const eventTarget = event.target as HTMLInputElement;
+    const file: File | undefined = eventTarget.files?.[0];
     if (file) {
       const formData: FormData = new FormData();
       formData.append('file', file);
